Make components just for Items.
This reduces the number of conditionals in the views, making them easier to understand
