+++ app/components/show/item_component.html.erb
@@ -22,10 +22,10 @@
 
 <div class="row mt-5">
   <div class="col-md-6">
-    <%= render Show::OverviewComponent.new(presenter: presenter) %>
+    <%= render Show::Item::OverviewComponent.new(presenter: presenter) %>
   </div>
   <div class="col-md-6">
-    <%= render Show::DetailsComponent.new(presenter: presenter) %>
+    <%= render Show::Item::DetailsComponent.new(presenter: presenter) %>
   </div>
 </div>
