--- app/components/show/overview_component.html.erb
@@ -1,66 +1,0 @@
-<table class="table table-sm mb-5 caption-header">
-    <caption>Overview</caption>
-    <tbody>
-      <tr>
-        <th class="col-3" scope="row">DRUID</th>
-        <td>
-            <%= id %>
-        </td>
-      </tr>
-
-      <% unless admin_policy? %>
-        <tr>
-          <th class="col-3" scope="row">Admin policy</th>
-          <td>
-            <%= render Show::GovernedByComponent.new(document: @solr_document, state_service: state_service) %>
-          </td>
-        </tr>
-      <% end %>
-
-      <% if item? %>
-        <tr>
-          <th class="col-3" scope="row">Collection</th>
-          <td>
-            <%= render Show::IsMemberOfComponent.new(document: @solr_document, state_service: state_service) %>
-          </td>
-        </tr>
-      <% end %>
-
-      <tr>
-        <th class="col-3" scope="row">Status</th>
-        <td>
-            <%= status %>
-        </td>
-      </tr>
-
-      <tr>
-        <th class="col-3" scope="row">Access rights</th>
-        <td>
-          <%= render Show::AccessRightsComponent.new(document: @solr_document, state_service: state_service) %>
-        </td>
-      </tr>
-
-      <% if item? || collection? %>
-        <tr>
-          <th class="col-3" scope="row">Copyright</th>
-          <td>
-            <%= render Show::CopyrightComponent.new(change_set: @presenter.change_set, state_service: state_service) %>
-          </td>
-        </tr>
-
-        <tr>
-          <th class="col-3" scope="row">License</th>
-          <td>
-            <%= render Show::LicenseComponent.new(change_set: @presenter.change_set, state_service: state_service) %>
-          </td>
-        </tr>
-
-        <tr>
-          <th class="col-3" scope="row">Use and reproduction</th>
-          <td>
-            <%= render Show::UseStatementComponent.new(change_set: @presenter.change_set, state_service: state_service) %>
-          </td>
-        </tr>
-      <% end %>
-    </tbody>
-  </table>
