+++ spec/components/show/apo/details_component_spec.rb
@@ -3,13 +3,22 @@
 require 'rails_helper'
 
 RSpec.describe Show::Apo::DetailsComponent, type: :component do
-  pending "add some examples to (or delete) #{__FILE__}"
+  let(:component) { described_class.new(presenter: presenter) }
+  let(:presenter) { instance_double(ArgoShowPresenter, document: doc, cocina: cocina, state_service: state_service) }
+  let(:cocina) { instance_double(Cocina::Models::AdminPolicy) }
+  let(:rendered) { render_inline(component) }
+  let(:allows_modification) { true }
+  let(:state_service) { instance_double(StateService, allows_modification?: allows_modification) }
+  let(:doc) do
+    SolrDocument.new('id' => 'druid:kv840xx0000',
+                     SolrDocument::FIELD_REGISTERED_DATE => ['2012-04-05T01:00:04.148Z'],
+                     SolrDocument::FIELD_OBJECT_TYPE => object_type)
+  end
+  let(:object_type) { 'adminPolicy' }
 
-  # it "renders something useful" do
-  #   expect(
-  #     render_inline(described_class.new(attr: "value")) { "Hello, components!" }.css("p").to_html
-  #   ).to include(
-  #     "Hello, components!"
-  #   )
-  # end
+  context 'when allows_modification is true' do
+    it 'renders the appropriate buttons' do
+      expect(rendered.css("a[aria-label='Edit tags']")).to be_present
+    end
+  end
 end
